requirements.txt
fastapi
uvicorn
pymupdf
python-docx
google-generativeai
openai-whisper
opencv-python
selenium
python-multipart